% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{build_test_data}
\alias{build_test_data}
\title{Create unit tests based on a workflow step for an operator}
\usage{
build_test_data(
  out_table,
  ctx,
  test_name,
  test_folder = NULL,
  version = "",
  absTol = NULL,
  relTol = NULL,
  r2 = NULL,
  docIdMapping = c(),
  props = c()
)
}
\arguments{
\item{out_table}{operator output table with namespace}

\item{ctx}{Tercen context.}

\item{test_name}{Name identifying the test.}

\item{test_folder}{Path where test files are stored.}

\item{version}{Corresponding operator version when test was created.}

\item{absTol}{Absolute tolerance parameter.}

\item{relTol}{Relative tolerance parameter.}

\item{r2}{Squared correlation parameter.}

\item{docIdMapping}{mapping between document Id and filename used for testing.}

\item{props}{Array containing properties to be set for specific stepIds.}
}
\description{
Based on the output table and input projection from an operator, builds
the necessary files for a test case.
}
\examples{
build_test_data( tbl, ctx, paste0(step_name, "_absTol"),
version = '0.0.1',
absTol = 0.001)

build_test_data( tbl, ctx, paste0(step_name, "_absTol"),
version = '0.0.1',
absTol = 0.001,
docIdMapping = c("32d4b2986b98f3ebd5b5baa990000148"="hospitals.csv.zip"))
}
\keyword{test}
