% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_utils.R
\name{check_ctx}
\alias{check_ctx}
\title{Check for the presence of inputs.}
\usage{
check_ctx(ctx, what, min_n = 1)
}
\arguments{
\item{ctx}{Tercen context.}

\item{what}{Which factor should be checked? Any of x, .x, y, .y, row, rows,
column, columns, colors, labels}
}
\description{
This function checks whether the specified list of inputs has been populated.
}
\examples{
ctx <- tercen::tercenCtx()
check_ctx(ctx, c("colors"))
}
\keyword{utils}
