% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{check_test_local}
\alias{check_test_local}
\title{Check an operator result against a previously saved run}
\usage{
check_test_local(
  out_table,
  test_name,
  test_folder = NULL,
  metric = "eq",
  absTol = 0,
  relTol = 0,
  r2 = 0
)
}
\arguments{
\item{out_table}{operator output table with namespace}

\item{test_name}{Name identifying the test.}

\item{metric}{Which comparison to use.}

\item{absTol}{Absolute tolerance parameter.}

\item{relTol}{Relative tolerance parameter.}

\item{r2}{Squared correlation parameter.}
}
\description{
Base don a test name, compares a computed table agianst a previously saved one.
}
\examples{
check_test_local(tbl, paste0(step_name, "_absTol"),
absTol = 0.001)

}
\keyword{test}
