% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HS.R
\name{estimate_hs}
\alias{estimate_hs}
\title{Estimates the model of Huang & Stoll (1997)}
\usage{
estimate_hs(price_diff, price, indicator, indicator_lag, midquote)
}
\arguments{
\item{price_diff}{a numeric vector containing the series of first price differences.}

\item{price}{a numeric vector containing the price series.}

\item{indicator}{an integer vector containing the trade direction with a buy as 1 and
a sell as -1.}

\item{indicator_lag}{an integer vector containing the first lag of the indicator series.}

\item{midquote}{a numeric vector with the midquote price series.}
}
\value{
A data.frame with the following values:
\describe{
 \item{n}{the number of observation used in estimation.}
 \item{beta1}{the effective half-spread.}
 \item{beta1_std}{the standard deviation of the effective half-spread.}
 \item{beta1_t}{the t-value of the effective half-spread.}
 \item{beta1_p}{the p-value of the effective half-spread.}
 \item{beta2}{the part of the spread due to adverse selection and inventory.}
 \item{beta2_std}{the standard deviation of the adverse selection and inventory\cr
 spread component.}
 \item{beta2_t}{the t-value of the adverse selection and inventory\cr 
 spread component.}
 \item{beta2_p}{the p-value of the adverse selection and inventory\cr 
 spread component.}
 \item{r2}{the coefficient of determination.}
 \item{r2_adj}{the adjusted coefficient of determination.}
 \item{f_stat}{the value of the F-statistic.}
 \item{f_pval}{the p-value of the F-statistic.}
 \item{spread_eff_est}{the effective spread estimated from the model using the 
 formula \eqn{2(\theta+\phi)}.}
 \item{spread_eff_std}{the standard deviation of the effective spread, calculated 
 via the \emph{delta method}.}
 \item{spread_eff_emp}{the empirical effective spread, calculated directly from the data 
 by the arithmetic mean of the series \eqn{q_t(p_t-m_t)}, where \eqn{q_t} is the trade 
 direction, \eqn{p_t} the transaction price, and \eqn{m_t} the price midquote series.}
 \item{spread_eff_emp_std}{the standard deviation of the empirical effective spread calculated 
 as the standard deviation of the series \eqn{q_t(p_t-m_t)}.}
 \item{spread_eff_emp_se}{the standard error of the estimated empirical effective spread 
 using the formula \eqn{SE=STD/\sqrt n}.}
 \item{spread_eff_emp_med}{the median of the empirical effective spread calculated as the 
 median of the series \eqn{q_t(p_t-m_t)}.}
}
}
\description{
Estimates the model of Huang & Stoll (1997)
}
\details{
The function estimates for given data the trade indicator model of 
\href{https://academic.oup.com/rfs/article-abstract/10/4/995/1605656}{Huang & Stoll (1997)}. 
For estimation an OLS approach is used similar to the one desribed in the paper of the two 
authors. For application the \code{\link{lm}}-function is used together with
\code{\link[sandwich]{NeweyWest}} for NeweyWest standard errors. For details
it is referred to the \code{\link[sandwich]{NeweyWest}}-function.
}
\references{
Huang & Stoll (1997), "The Component of the Bif-Ask Spread: A 
General Approach," The Review of Financial Studies, Vol. 10, Issue 4., pp. 995-1034.
}
