% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HS.R
\name{estimate_hs_mod}
\alias{estimate_hs_mod}
\title{Estimates the model of Huang & Stoll (1997) using the approach by 
Theissen & Zehnder (2015)}
\usage{
estimate_hs_mod(price_diff, price, indicator, indicator_lag, midquote,
  midquote_lag)
}
\arguments{
\item{price_diff}{a numeric vector containing the series of first price differences.}

\item{price}{a numeric vector containing the price series.}

\item{indicator}{an integer vector containing the trade direction with a buy as 1 and
a sell as -1.}

\item{indicator_lag}{an integer vector containing the first lag of the trade indicator series.}

\item{midquote}{a numeric vector with the midquote price series.}

\item{midquote_lag}{a numeric vector containing the first lag of the midquote series.}
}
\value{
A data.frame with the following values:
\describe{
 \item{n}{the number of observation used in estimation.}
 \item{a}{the first parameter estimate of the OLS model in step 1, \eqn{\alpha}.}
 \item{a_std}{the standard deviation of the first parameter estimate of the OLS 
 model in step 1, \eqn{\alpha}.}
 \item{a_t}{the t-value of the first parameter estimate of the OLS  model in step 1, 
 \eqn{\alpha}.}
 \item{a_pval}{the p-value of the first parameter estimate of the OLS  model in step 1, 
 \eqn{\alpha}.}
 \item{psi12}{the parameter (1,2) of the VARMA model in step 2.}
 \item{psi12_std}{the standard deviation of the parameter (1,2) of the VARMA model 
 in step 2.}
 \item{psi12_tval}{the t-value of the parameter (1,2) of the VARMA model in step 3.}
 \item{psi12_pval}{the p-value of the paramater (1,2) of the VARMA model in step 3.}
 \item{varv}{the variance of the trade innovation, \eqn{\nu} from step 1.}
 \item{varu}{the variance of the public information arrival, \eqn{u} from step 2.}
 \item{covuv}{the covariance of the trade innovation, \eqn{\nu} from step 1, and
 the public information arrival, \eqn{u} from step 2.}
 \item{spread_eff_est1}{the estimation of the effective spread using the traditional 
 estimation approach.}
 \item{spread_eff_est2}{the estimation of the effective spread using the modified 
 estimation approach.} 
 \item{spread_eff_emp}{the empirical effective spread measured from the data.}
 \item{spread_eff_emp_std}{the standard deviation of the empirical effective spread calculated 
 as the standard deviation of the series \eqn{q_t(p_t-m_t)}.}
 \item{spread_eff_emp_se}{the standard error of the estimated empirical effective spread 
 using the formula \eqn{SE=STD/\sqrt n}.}
 \item{spread_eff_emp_med}{the median of the empirical effective spread calculated as the 
 median of the series \eqn{q_t(p_t-m_t)}.}
 \item{stable}{the stability of the estimated polynomials, i.e. if all roots are 
 inside the unit circle.}
}
}
\description{
Estimates the model of Huang & Stoll (1997) using the approach by 
Theissen & Zehnder (2015)
}
\details{
The function estimates for given data the trade indicator model of 
\href{https://academic.oup.com/rfs/article-abstract/10/4/995/1605656}{Huang & 
Stoll (1997)} using the modified estimation approach by 
Theissen & Zehnder (2015). The latter authors use a two-step estimation approach 
to overcome the negative bias in the estimation of the effective spread. The major
part of this bias appears to come from the adverse selection component \eqn{\theta}
in the trade indicator model of Huang & Stoll (HS). 

The authors propose a two-step estimation procedure that results in a less biased
estimation of both, the effective spread and the adverse selection component, as they
explicitly account for endogeneity in their model, namely a correlation between the
trade innovation and the public information arrival. The estimation procedure involves
a VARMA part that is estimated using the \code{\link{dse}}-package.

Standard deviations for the VARMA model (step 2) are estimated using the \emph{delta 
method}.
}
\references{
Huang & Stoll (1997), "The Component of the Bif-Ask Spread: A 
General Approach," The Review of Financial Studies, Vol. 10, Issue 4., pp. 995-1034.

Zehnder & Theissen (2015), "Estimation of Trading Costs: Trade Indicator Models 
Revisited," unpublished.
}
