% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRR.R
\name{estimate_mrr}
\alias{estimate_mrr}
\title{Estimates the model of Madhavan, Richardson, and Roomans (1997)}
\usage{
estimate_mrr(price_diff, price, indicator, indicator_lag, midquote,
  midquote_lag, spread_quoted)
}
\arguments{
\item{price_diff}{a numeric vector containing the series of first price differences.}

\item{price}{a numeric vector containing the price series.}

\item{indicator}{an integer vector containing the trade direction with a buy as 1 and
a sell as -1.}

\item{indicator_lag}{an integer vector containing the first lag of the indicator series.}

\item{midquote}{a numeric vector with the midquote price series.}

\item{midquote_lag}{a numeric vector containing the first lag of the midquote series.}

\item{spread_quoted}{a numeric vector containing the quoted spread series.}
}
\value{
A data.frame with the following values:
\describe{
 \item{n}{the number of observation used in estimation.}
 \item{theta}{the adverse selection component.}
 \item{theta_std}{the standard deviation of the adverse selection component.}
 \item{theta_t}{the t-value of the adverse selection component.}
 \item{theta_p}{the p-value of the adverse selection component.}
 \item{phi}{the transitory cost component.}
 \item{phi_std}{the standard deviation of the transitory cost component.}
 \item{phi_t}{the t-value of the transitory cost component.}
 \item{phi_p}{the p-value of the transitory cost component.}
 \item{rho}{the indicator first order autocorrelation.}
 \item{rho_std}{the standard deviation of the indicator autocorrelation.}
 \item{rho_t}{the t-value of the indicator autocorrelation.}
 \item{rho_p}{the p-value of the indicator autocorrelation.}
 \item{r2}{the coefficient of determination.}
 \item{r2_adj}{the adjusted coefficient of determination.}
 \item{f_test}{the value of the F-statistic.}
 \item{f_pval}{the p-value of the F-statistic.}
 \item{theta_start}{the start value for theta in the numerical optimization.}
 \item{phi_start}{the start value for phi in the numerical optimization.}
 \item{rho_start}{the start value for rho in the numerical optimization.}
 \item{eps_std}{the estimated standard deviation of the epsilon error term 
 (order flow innovations).}
 \item{eta_std}{the estimated standard deviation of the eta error term 
 (stochastic rounding errors).}
 \item{spread_eff}{the effective spread estimated from the model using the 
 formula \eqn{2(\theta+\phi)}.}
 \item{spread_eff_std}{the standard deviation of the effective spread, calculated 
 via the \emph{delta method}.}
 \item{spread_eff_emp}{the empirical effective spread, calculated directly from the data 
 by the arithmetic mean of the series \eqn{q_t(p_t-m_t)}, where \eqn{q_t} is the trade 
 direction, \eqn{p_t} the transaction price, and \eqn{m_t} the price midquote series.}
 \item{spread_eff_emp_std}{the standard deviation of the empirical effective spread calculated 
 as the standard deviation of the series \eqn{q_t(p_t-m_t)}.}
 \item{spread_eff_emp_se}{the standard error of the estimated empirical effective spread 
 using the formula \eqn{SE=STD/\sqrt n}.}
 \item{spread_eff_emp_med}{the median of the empirical effective spread calculated as the 
 median of the series \eqn{q_t(p_t-m_t)}.}
 \item{spread_quoted}{the mean quoted spread using the input quoted spread series.}
 \item{spread_quoted_std}{the standard deviation of the quoted spread using the input 
 quoted spread series.}
 \item{spread_quoted_se}{the standard error of the quoted spread estimate using the 
 formula \eqn{SE=STD/\sqrt n}.}
 \item{spread_quoted_med}{the median of the quoted spread series.}
}
}
\description{
Estimates the model of Madhavan, Richardson, and Roomans (1997)
}
\details{
The function estimates for given data the trade indicator model of 
\href{https://academic.oup.com/rfs/article-abstract/10/4/1035/1605749}{Madhavan, 
Richardson, and Roomans (1997)}. For estimation a GMM approach is used similar to 
the one desribed in the paper of the three authors. For application the \code{\link{gmm}}-
package is used with a \emph{Bartlett}-kernel, a heteroscedastic and autocorrelation 
consistent variance-covariance matrix with a \emph{Newey West bandwith}. For details
it is referred to the \code{\link[gmm]{gmm}}-function.
}
\references{
Madhavan, Richardson & Roomans (1997), "Why Do Security Prices Change? A
Transaction-Level Analysis of NYSE Stocks," Review of Financial Studies, Vol. 10, 
No. 4, pp. 1035-1064.
}
