% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRR.R
\name{estimate_mrr_mod}
\alias{estimate_mrr_mod}
\title{Estimates the model of Madhavan, Richardson & Roomans (1997) using the approach by 
Theissen & Zehnder (2015)}
\usage{
estimate_mrr_mod(price_diff, price, indicator, indicator_lag, indicator_lag2,
  midquote, midquote_lag)
}
\arguments{
\item{price_diff}{a numeric vector containing the series of first price differences.}

\item{price}{a numeric vector containing the price series.}

\item{indicator}{an integer vector containing the trade direction with a buy as 1 and
a sell as -1.}

\item{indicator_lag}{an integer vector containing the first lag of the trade indicator series.}

\item{indicator_lag2}{an integer vector containing the second lag of the trade indicator series.}

\item{midquote}{a numeric vector with the midquote price series.}

\item{midquote_lag}{a numeric vector containing the first lag of the midquote series.}
}
\value{
A data.frame with the following values:
\describe{
 \item{n}{the number of observation used in estimation.}
 \item{rho}{the estimated first order autocorrelation, \eqn{\rho}, of the trade 
 indicator series, i.e. the parameter of the AR(1) model in step 1.}
 \item{rho_std}{the standard deviation of the estimated first order autocorrelation,
 \eqn{\rho} of step 1.}
 \item{a1}{the first parameter estimate of the OLS model in step 2, \eqn{\alpha_1}.}
 \item{a1_std}{the standard deviation of the first parameter estimate of the OLS 
 model in step 2, \eqn{\alpha_1}.}
 \item{a2}{the second parameter estimate of the OLS model in step 2, \eqn{\alpha_2}.}
 \item{a2_std}{the standard deviation of the second parameter estimate of the OLS 
 model in step 2, \eqn{\alpha_2}.}
 \item{psi1}{the parameter (1,2) of the VARMA model in step 3.}
 \item{psi1_std}{the standard deviation of the parameter (1,2) of the VARMA model 
 in step 3.}
 \item{psi2}{the parameter (2,2) of the VARMA model in step 3.}
 \item{psi2_std}{the standard deviation of the parameter (2,2) of the VARMA model 
 in step 3.}
 \item{varv}{the variance of the trade innovation, \eqn{\nu} from step 1.}
 \item{varu}{the variance of the public information arrival, \eqn{u} from step 2.}
 \item{covuv}{the covariance of the trade innovation, \eqn{\nu} from step 1, and
 the public information arrival, \eqn{u} from step 2.}
 \item{spread_eff_emp}{the empirical effective spread measured from the data.}
 \item{spread_eff_est1}{the estimation of the effective spread using the traditional 
 estimation approach.}
 \item{spread_eff_est2}{the estimation of the effective spread using the modified 
 estimation approach.} 
 \item{stable}{the stability of the estimated polynomials, i.e. if all roots are 
 inside the unit circle.}
}
}
\description{
Estimates the model of Madhavan, Richardson & Roomans (1997) using the approach by 
Theissen & Zehnder (2015)
}
\details{
The function estimates for given data the trade indicator model of 
\href{https://academic.oup.com/rfs/article-abstract/10/4/1035/1605749}{Madhavan, 
Richardson & Roomans (1997)} using the modified estimation approach by 
Theissen & Zehnder (2015). The latter authors use a three-step estimation approach 
to overcome the negative bias in the estimation of the effective spread. The major
part of this bias appears to come from the adverse selection component \eqn{\theta}
in the trade indicator model of Madhavan, Richardson & Roomans (MRR). 

The authors propose a three-step estimation procedure that results in an unbiased
estimation of both effective spread and also the adverse selection component as they
explicitly account for endogeneity in their model, namely a correlation between the
trade innovation and the public information arrival. The estimation procedure involves
a VARMA part that is estimated using the \code{\link{dse}}-package.

Standard deviations for the VARMA model (step 3) are estimated using the \emph{delta 
method}.
}
\references{
Madhavan, Richardson & Roomans (1997), "Why Do Security Prices Change? A 
Transaction-Level Analysis of NYSE Stocks," Review of Financial Studies, Vol. 10, 
No. 4, pp. 1035-1064.

Zehnder & Theissen (2015), "Estimation of Trading Costs: Trade Indicator Models 
Revisited," unpublished.
}
