% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialisation_utils.R
\name{find_schema_by_factor_name}
\alias{find_schema_by_factor_name}
\title{Retrieve serialized object in tercen}
\usage{
find_schema_by_factor_name(ctx, factor.name)
}
\arguments{
\item{ctx}{Tercen context.}

\item{factor_name}{Name of the factor.}
}
\description{
This function finds the schema containing the factor name.
}
\keyword{utils}
