% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialisation_utils.R
\name{get_serialized_result}
\alias{get_serialized_result}
\title{Get a serialized tercen result}
\usage{
get_serialized_result(df, object, object_name, ctx)
}
\arguments{
\item{df}{Data frame containing results.}

\item{object}{Object to be serialized.}

\item{object_name}{Name of the object.}

\item{ctx}{Tercen context.}
}
\description{
This function allows you to prepare an object to be sent back to Tercen.
}
\examples{
str <- serialise_to_string(iris)
iris2 <- deserialise_from_string(str)
}
\keyword{utils}
