% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{get_step_name}
\alias{get_step_name}
\title{Utility function to extract the step name given an workflow and step IDs}
\usage{
get_step_name(ctx, wkfId, stepId)
}
\arguments{
\item{ctx}{Tercen context.}

\item{wkfId}{Workflow Identifier}

\item{stepId}{Step Identifier}
}
\description{
Utility function to extract the step name given an workflow and step IDs
}
\keyword{test}
