% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{Hello world}
\usage{
hello_world()
}
\arguments{
\item{none}{No parameters.}
}
\description{
This function allows you to print hello world.
}
\examples{
hello_world()
}
\keyword{test}
