% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lee_ready_vec}
\alias{lee_ready_vec}
\title{Applies the Lee & Ready (1991) algorithm}
\usage{
lee_ready_vec(price, bid_price, ask_price)
}
\arguments{
\item{price}{a numeric vector containing the transaction price series.}

\item{bid_price}{a numeric vector containing the bid price series.}

\item{ask_price}{a numeric vector containing the ask price series.}
}
\value{
an integer vector containing the trade direction with a buy 
indicated as 1 and a sell indicated as -1
}
\description{
Applies the Lee & Ready (1991) algorithm
}
\details{
This function applies the algorithm by 
\href{https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1540-6261.1991.tb02683.x}{Lee & Ready (1991)}
to infer trade direction from prices. The function runs fast C++-code in the 
background using the \code{\link{Rcpp}} functionality. Note that the 
length of all input vectors have to be the same, otherwise the C++-code 
throws an exception.
}
\references{
Lee and Ready (1991), "Inferring Trade Direction from Intraday Data," 
Journal of Finance, Vol. 42, Issue 2
}
