% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{load_data}
\alias{load_data}
\title{Downloads and unzip files with cache support.}
\usage{
load_data(ctx, documentId, force_load = FALSE)
}
\arguments{
\item{ctx}{Tercen context.}

\item{documentId}{Identifier for file to be loaded.}

\item{force_load}{Downloads and unzips data regardless of cache}
}
\description{
Searches the /tmp/documentId folder for flags indicating that the data
associated with the documentId has previously been loaded.
If it has not, downloads the file and, if it is a ZIP archive, unzips it.
The cache is cleared by the GarbageCollector
}
\examples{
load_data(ctx, docId)
}
\keyword{file}
