% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_utils.R
\name{matrix_to_flowFrame}
\alias{matrix_to_flowFrame}
\title{Convert a matrix into a flowFrame object.}
\usage{
matrix_to_flowFrame(data)
}
\arguments{
\item{object}{Object to be serialized.}
}
\description{
This function allows you to convert a matrix obtained from a tercen context
into a flowFrame object.
}
\examples{
matrix_to_flowFrame(data)
}
\keyword{utils}
