% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nocb_1min_df}
\alias{nocb_1min_df}
\title{Calculates the 1-minute price lead and corresponding timestamp}
\usage{
nocb_1min_df(price, time_stamp)
}
\arguments{
\item{price}{a numeric vector containing the price series.}

\item{tradetime}{a POSIXct vector containing the time stamps of the price series.}
}
\value{
a data.frame containing the 1-minute price lead and corresponding timestamp.
}
\description{
Calculates the 1-minute price lead and corresponding timestamp
}
\details{
This function calculates the 1-minute lead price corresponding to each 
of the entries in \code{price}. The return value has the same length as the input 
vector to ensure compatibility inside of e.g. a \code{data.frame}. 

The \code{\link{POSIXct}} datetime format is a requirement as this function runs 
fast C++-code in the background that expects a double format which is provided by
the \code{POSIXct} format  (see herefore the 
\href{http://dirk.eddelbuettel.com/code/rcpp/html/classRcpp_1_1Datetime.html}{Rcpp::Datetime}). 

For matching an NOCB (next observation carried backwards) approach is used, i.e. 
if for a price with certain timestamp there exists 1-minute price lead, the next 
>1-minute price lead is used. This leads possibly to repeating price lead entries. 
Furthermore, as the length of the return vector is the same as the input vector,
the last entries of the return vector are usually \code{NA} values.

If no timestamp is required it is referred to the \code{vector}-equivalent 
\code{\link{nocb_1min_vec}}, which returns next to the 1-minute lead price 
series in addition the 1-minute lead timestamp.
}
\seealso{
\code{\link{nocb_1min_vec}}, \code{\link{nocb_5min_vec}}, \code{\link{nocb_5min_df}}
}
