% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nocb_5min_vec}
\alias{nocb_5min_vec}
\title{Calculates the 5-minute price lead}
\usage{
nocb_5min_vec(price, time_stamp)
}
\arguments{
\item{price}{a numeric vector containing the price series.}

\item{tradetime}{a POSIXct vector containing the time stamps of the price series.}
}
\value{
a numeric vector containing the 5-minute price lead.
}
\description{
Calculates the 5-minute price lead
}
\details{
This function calculates the 5-minute lead price corresponding to each 
of the entries in \code{price}. The return value has the same length as the input 
vector to ensure compatibility inside of e.g. a \code{data.frame}. 

The \code{\link{POSIXct}} datetime format is a requirement as this function runs 
fast C++-code in the background that expects a double format which is provided by
the \code{POSIXct} format  (see herefore the 
\href{http://dirk.eddelbuettel.com/code/rcpp/html/classRcpp_1_1Datetime.html}{Rcpp::Datetime}). 

For matching an NOCB (next observation carried backwards) approach is used, i.e. 
if for a price with certain timestamp there exists 5-minute price lead, the next 
>5-minute price lead is used. This leads possibly to repeating price lead entries. 
Furthermore, as the length of the return vector is the same as the input vector,
the last entries of the return vector are usually \code{NA} values.

If a timestamp is required for reasons of documentation or testing 
it is referred to the \code{data.frame}-equivalent \code{\link{nocb_5min_df}}, 
which returns next to the 5-minute lead price series in addition the 5-minute 
lead timestamp.
}
\seealso{
\code{\link{nocb_5min_df}}, \code{\link{nocb_1min_vec}}, \code{\link{nocb_1min_df}}
}
