% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialisation_utils.R
\name{plot_file_to_df}
\alias{plot_file_to_df}
\title{Plot file (PNG, SVG, PDF) to data frame.}
\usage{
plot_file_to_df(file_path, filename = NULL)
}
\arguments{
\item{file_path}{A vector containing paths to PNG files.}

\item{filename}{(optional) A vector containing output file names.}
}
\description{
This function finds the schema containing the factor name.
}
\keyword{utils}
