% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialisation_utils.R
\name{png_to_df}
\alias{png_to_df}
\title{PNG to data frame.}
\usage{
png_to_df(png_file_path, filename = NULL)
}
\arguments{
\item{png_file_path}{A vector containing paths to PNG files.}

\item{filename}{(optional) A vector containing output file names.}
}
\description{
This function finds the schema containing the factor name.
}
\keyword{utils}
