% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_utils.R
\name{populate_gh_workflow}
\alias{populate_gh_workflow}
\title{Populate GitHub workflows}
\usage{
populate_gh_workflow(type, output_path = "./.github/workflows")
}
\arguments{
\item{path}{Output path (default).}
}
\description{
Based on the output table and input projection from an operator, builds
the necessary files for a test case.
}
\examples{
populate_gh_workflow(type = "docker")
}
\keyword{test}
