% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{price_impact}
\alias{price_impact}
\title{Compute price impact}
\usage{
price_impact(indicator, price, price_lead)
}
\arguments{
\item{indicator}{an integer vector or numeric vector for \eqn{t}.}

\item{price}{a numeric vector holding the prices for \eqn{t}.}

\item{price_lead}{a numeric vector holding the leading prices for \eqn{t+x}.}
}
\description{
\code{price_impact} calculates the price impact for a certain lead \eqn{x}.
}
\details{
This function takes intraday data from a financial exchange and calculates the 
price impact defined \eqn{ PI = \frac{1}{T}\sum_{t=1}^Tq_t(m_{t+x}-m_t) }, where 
\eqn{ q_t } is the price indicator at time \eqn{ t }, \eqn{ m_t } is the midquote 
at time\eqn{ t } and \eqn{ x } is a certain time interval, e.g. 1 or 5 minutes. 
So, the indicator time always corresponds to the starting midquote \eqn{ m_t }. 
All price impacts are calculated in US $-Cents.

The function uses internally makes use of \code{\link{mean}} and excluding all 
NA values (making use of its input argument \code{na.rm = TRUE}). Note that 
the user has to prepare the time series data, i.e. calculate for the \code{price} 
data with index \eqn{t} the corresponding \code{price_lead} with index \eqn{t+x}.
}
