% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{run_local_test}
\alias{run_local_test}
\title{Check an operator result against a previously saved run}
\usage{
run_local_test(
  res_table,
  ctx,
  test_name,
  test_folder = NULL,
  absTol = NULL,
  relTol = NULL,
  r2 = NULL,
  skipCols = c(),
  forcejoincol = NULL,
  forcejoinrow = NULL
)
}
\arguments{
\item{res_table}{operator output table with namespace}

\item{ctx}{Tercen context}

\item{test_name}{Name identifying the test.}

\item{absTol}{Absolute tolerance parameter.}

\item{relTol}{Relative tolerance parameter.}

\item{r2}{Squared correlation parameter.}

\item{skipCols}{Column names to be skipped during testing}

\item{forcejoincol}{Logical array matching num. of out_table to force relation to column (when join was not made through .ci).}

\item{forcejoinrow}{Logical array matching num. of out_table to force relation to row (when join was not made through .ri).}
}
\description{
Based on a test name, compares a computed table against a previously saved one.
}
\examples{
check_test_local(tbl, ctx, paste0(step_name, "_absTol"), absTol = 0.001)

}
\keyword{test}
