% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialisation_utils.R
\name{save_plot}
\alias{save_plot}
\title{Save plot as a temporary file.}
\usage{
save_plot(plt, type = "png", ...)
}
\arguments{
\item{plt}{ggplot object.}

\item{type}{any of png, pdf, svg.}
}
\description{
Save plot as a temporary file.
}
\keyword{utils}
