% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialisation_utils.R
\name{serialise_to_string}
\alias{serialise_to_string}
\title{Serialise an object.}
\usage{
serialise_to_string(object)
}
\arguments{
\item{object}{Object to be serialised.}
}
\description{
This function allows you to serialise an object to a string.
}
\examples{
serialise_to_string(iris)
}
\keyword{utils}
