% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialisation_utils.R
\name{serialize_to_string}
\alias{serialize_to_string}
\title{Serialize an object.}
\usage{
serialize_to_string(object)
}
\arguments{
\item{object}{Object to be serialized.}
}
\description{
This function allows you to serialize an object to a string.
}
\examples{
serialize_to_string(iris)
}
\keyword{utils}
