% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_utils.R
\name{set_tercen_credentials}
\alias{set_tercen_credentials}
\title{Set Tercen credentials.}
\usage{
set_tercen_credentials(force = FALSE)
}
\arguments{
\item{force}{Whether to force credentials update or not (default: FALSE).}
}
\description{
Set Tercen credentials.
}
