% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_utils.R
\name{set_workflow_step_ids}
\alias{set_workflow_step_ids}
\title{Set workflow and data step IDs from URL.}
\usage{
set_workflow_step_ids(step_url, set_credentials = TRUE, serviceUri = NULL)
}
\arguments{
\item{step_url}{Data step URL.}

\item{set_credentials}{Whether to set credentials (default is TRUE).}

\item{serviceUri}{Service URI.}
}
\description{
Set workflow and data step IDs from URL.
}
