\name{tim-package}
\alias{tim-package}
\alias{tim}
\docType{package}
\title{
\packageTitle{tim}
}
\description{
\packageDescription{tim}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tim}
\packageIndices{tim}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{tim}

Maintainer: \packageMaintainer{tim}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
